// src/shared/infraestructure/enums/message.enum.ts
export enum MessageEnum {
  SERVICE_STARTED = 'Service started',
  NOT_FUNDS_LOCAL = 'There are not enough funds in the local account',
  NOT_FUNDS_EXCHANGE = 'There are not enough funds in the exchange account',
  TRANSACTION_NOT_PROCESSED = 'Transaction not processed correctly',
  TRANSACTION_PROCESSED = 'Transaction processed correctly',
  TRANSACTION_CANCELED = 'Transaction canceled',
  ENTITY_SELECT = 'Data returned successfully.'
}